StateSubtasks & resetting
Decided to break down tasks logic into an internal subtask to more easily reuse and modify logic in states

Added reset option to shared values as most use cases they are expected to set their values at runtime. However, this is more of a temporary solution. I didnt just not serialized the data simply to make it easier to debug for now