Attacking WIP & Separating States
Started working on implementing a prototype for attacking an enemy.

Started separating the old select target state into two separate states:
- Select State: Exclusively for selecting a target
- Move to target: Exclusively to move to a selected target

For this I added a new shared data reference in the blackboard for states to keep a reference to the selected unit target in preparation of the attacking state.

Some renaming that was missed was also included