Damaged animation & refactoring
Changed GameEventListenerComponent to be able to store several listeners so that there wont be a need to include a new component for each new listener.

Updated animation trigger parameter setter to only require one parameter name as having an extra variable and one that was not being used was confusing and not intuitive.