Directional Animations & Rework
Had to refactor and reorganise the structure of Animator Data Setter class, formerly named Animation Data Setter to accommodate for double parameter setting.

It also enable me to start fixing some of the repetition issues, however it does mean that unless a custom editor is added for dynamic data setters, their parameter setters are being shown in the inspector even though they should be hidden, this is due to the parent class having them exposed.

The animations are still WIP but most of the initial work has been completed for the following animations, as the animation state machine had to be redone to include blend trees:
- Movement
- Idle
- Death